Add delete role confirmation dialog, clear form-field in role-rights

+++ src/pages/admin/admin-setup/role-management/role-list/role-list.page.ts
@@ -128,10 +128,12 @@ export class RoleListPage implements OnInit {
      */
     delete(role_guid) {
         const dialogRef = this.dialog.open(DialogDeleteConfirmationPage, {
-            // width: '250px',
+            data: role_guid
         });
         dialogRef.afterClosed().subscribe(result => {
-            this.roleAPi.delete_role_profile(role_guid).subscribe(response => this.ngOnInit())
+            if (result === role_guid) {
+                this.roleAPi.delete_role_profile(role_guid).subscribe(response => this.ngOnInit())
+            }
         });
     }
